feat(main): Handle CTRL+D (EOF)
Previously on CTRL+D key press, the rah_writer thread used to exit with an
Exception but the program didn't terminate as the rah_reader thread was still
running.

This commit, handles the EOF in rah_writer thread and joins the rah_reader
thread immediately after the rah_writer thread joins.

Signed-off-by: shailparmar03 <shailparmar26@gmail.com>